+++ components/form/index.ts
@@ -1,0 +1,5 @@
+export { default as Button } from "./Button";
+export { default as Icon } from "./Icon";
+export { default as Input } from "./Input";
+export { default as Label } from "./Label";
+export { default as OutlineButton } from "./OutlineButton";
